WIP Experimenting with inlining and generics
This was an attempt to help with inlining generic functions called in hot loops.
The thought was: the decision of which version of a function to run does
not need to be made very often, but the compiler doesn't know that.
Even very low-level functions like Bit are generic. In theory, the
decision only needs to made at the highest level: what kind of IP is the
user asking about when they call something like PrefixSet.Encompasses?