Fix treatment of ::0/0 and 0.0.0.0/0. Fixes #29.
We refer to these as the "default routes" in their respective IP spaces.
Storing and retrieving the default routes was not previously supported.

A default route is naturally represented by the zero key. In netipds,
every tree has a root node with no entry and whose key is the zero key.

Some care is required to allow the root node to hold an entry in the
trie. For example, if a method removes an entry by returning nil, then
removing an entry at the root node would replace a user's pointer to a
tree with nil.

Another reason we did not support the default routes was that calling
key.ToPrefix on the zero key returned the zero netip.Prefix, which is an
invalid value by design. It turned out this was self-imposed:
key.ToPrefix previously handled the zero key explicitly as a special
case. By removing this special case, and just using the general-case
logic, we get the mapping we want (key{} -> default route).