Use return value of tree.remove in PrefixSetBuilder.Remove
This return value doesn't actually need to be used here but:
- Better to be explicit
- It's used by PrefixMapBuilder.Remove

Note that NOT using the return value was actually upholding an implicit
invariant: the root node of the tree should never be replaced; it should
always be the original, zero-key node. We need this property in order
for tree.nilOrEmptyRoot to work.

This invariant is now explicit, and the one place where a child node
could replace the root node (in tree.remove) is now handled explicitly.

TODO: remove this invariant and allow the root node to be replaced. This
will require an additional change: the empty tree will need to be
represented by nil. Overall I think this design will be simpler.