Property-based test for PrefixSet.Merge + bugfix
The new test merges two relatively large, randomized PrefixSets and
checks the merged result's contents. This test exposed the following
bug:

In tree.mergeTree, when we encounter a t and o that do not overlap, we
need to add o's entire subtree, not just o itself. Otherwise, child
entries of o are not included in the merged result.